# OpenAPI.yaml
openapi: 3.0.3
info:
  title: API Blog Comments
  version: "1.0"
  description: >
    HTTP API for managing blog posts, comments, and authenticated users.
    The contract reflects a production-oriented baseline with real persistence,
    JWT-based authentication, ownership rules, and interactive documentation support.
servers:
  - url: http://localhost:5000
    description: Local execution endpoint
...
paths:
...
  /api/auth/login:
    post:
      tags: [Authentication]
      summary: Authenticate user
      description: >
        Authenticates a user using username and password and returns a JWT token.
        In production-oriented deployments, credentials must be provisioned explicitly.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/LoginRequest'
            example:
              username: operador
              password: senha-segura-123
      responses:
        '200':
          description: Authentication successful
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request payload
        '401':
          description: Invalid credentials
        '429':
          description: Too many authentication requests
  /api/auth/register:
    post:
      tags: [Authentication]
      summary: Register user
      description: >
        Creates a new user, stores the password with Argon2id,
        and returns a JWT token ready for authenticated requests.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/RegisterRequest'
            example:
              username: new-author
              password: senha-segura-123
      responses:
        '201':
          description: User created successfully
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthResponse'
        '400':
          description: Invalid request payload
        '409':
          description: Username already in use
          content:
            application/problem+json:
              schema:
                $ref: '#/components/schemas/ProblemDetails'
              example:
                title: Username conflict
                detail: The informed username is already in use. Choose a different value and try again.
                status: 409
        '429':
          description: Too many authentication requests
  /api/auth/me:
    get:
      tags: [Authentication]
      summary: Get authenticated user profile
      description: Returns the profile of the currently authenticated user.
      security:
        - bearerAuth: []
      responses:
        '200':
          description: Authenticated user profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AuthenticatedUserProfile'
        '401':
          description: Unauthorized
        '429':
          description: Too many authentication requests
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
...
    LoginRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          example: operador
        password:
          type: string
          format: password
          example: senha-segura-123
    RegisterRequest:
      type: object
      required:
        - username
        - password
      properties:
        username:
          type: string
          minLength: 3
          maxLength: 100
          example: luciano
        password:
          type: string
          format: password
          minLength: 8
          maxLength: 100
          example: senha-segura-123
    AuthResponse:
      type: object
      required:
        - token
      properties:
        token:
          type: string
          description: JWT bearer token to be used in the Authorization header
          example: eyJhbGciOiJIUzI1NiIsInR5cCI6IkpXVCJ9...
    AuthenticatedUserProfile:
      type: object
      required:
        - id
        - username
        - role
      properties:
        id:
          type: integer
          example: 10
        username:
          type: string
          example: author-user
        role:
          type: string
          example: Author
    ProblemDetails:
      type: object
      properties:
        type:
          type: string
          nullable: true
          example: https://tools.ietf.org/html/rfc9110#section-15.5.10
        title:
          type: string
          nullable: true
          example: Username conflict
        status:
          type: integer
          nullable: true
          example: 409
        detail:
          type: string
          nullable: true
          example: The informed username is already in use. Choose a different value and try again.
        instance:
          type: string
          nullable: true